# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Sweet cash"
  },
  "basePath": "/api/",
...
  "paths": {
...
    "/api/v1/transactions_category": {
      "post": {
        "tags": [
          "References"
        ],
        "summary": "Добавляет новую категории операции",
        "description": "",
        "operationId": "addCategotyTransaction",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CategoryTransaction"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Новая операция создана",
            "schema": {
              "$ref": "#/definitions/TransactionCategoryDetail"
            }
          },
          "400": {
            "description": "Неверно сформированный запрос"
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "403": {
            "description": "Отсутствует доступ к выполнению метода"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/api/v1/transactions/categories": {
      "get": {
        "tags": [
          "References"
        ],
        "summary": "Отдает все категории операции",
        "description": "",
        "operationId": "getTransactionCategories",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TransactionCategoryDetail"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/api/v1/transactions_category/{transactions_category_id}": {
      "put": {
        "tags": [
          "References"
        ],
        "summary": "Обновляет категорию операции",
        "description": "",
        "operationId": "updateTransactionCategories",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transactions_category_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "pattern": "\\d+",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/TransactionCategoryDetail"
            }
          },
          "400": {
            "description": "Неверно сформированный запрос"
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Authorization": {
      "description": "",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "CategoryTransaction": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "format": "STRING",
          "example": "Расход",
          "description": "Наименование типа"
        },
        "parent_category_id": {
          "type": "integer",
          "format": "int64",
          "example": 1,
          "description": "Номер корневой категории"
        },
        "description": {
          "type": "string",
          "example": "Создал для себя",
          "description": "Описание типа"
        }
      }
    },
    "TransactionCategoryDetail": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1,
          "description": "Номер категории"
        },
        "parent_category_id": {
          "type": "integer",
          "format": "int64",
          "example": 1,
          "description": "Номер корневой категории"
        },
        "name": {
          "type": "string",
          "format": "STRING",
          "example": "Продукты",
          "description": "Наименование категории"
        },
        "description": {
          "type": "string",
          "example": "Создал для себя",
          "description": "Описание категории"
        }
      }
    }
  }
}